SVN r3175 (ogl_hack)

+++ client/sdl/i_sound.h
@@ -4,6 +4,7 @@
 // $Id$
 //
 // Copyright (C) 1993-1996 by id Software, Inc.
+// Copyright (C) 2006-2012 by The Odamex Team.
 //
 // This program is free software; you can redistribute it and/or
 // modify it under the terms of the GNU General Public License
@@ -55,7 +56,7 @@ void I_LoadSound (struct sfxinfo_struct *sfx);
 int
 I_StartSound
 (int			id,
- int			vol,
+ float			vol,
  int			sep,
  int			pitch,
  bool			loop);
